Adjusted data types in models to be strictly correct per spec, includ…

+++ ISOv4Plugin/Mappers/GridMapper.cs
@@ -61,14 +61,14 @@ namespace AgGateway.ADAPT.ISOv4Plugin.Mappers
             }
 
             FileInfo info = new FileInfo(Path.Combine(this.TaskDataPath, Path.ChangeExtension(grid.Filename,"bin")));
-            grid.Filelength = info.Length;
+            grid.Filelength = (uint)info.Length;
 
             grid.GridMinimumNorthPosition = Convert.ToDecimal(GetOriginY(adaptRx));
             grid.GridMinimumEastPosition = Convert.ToDecimal(GetOriginX(adaptRx));
             grid.GridCellNorthSize = GetCellHeight(adaptRx);
             grid.GridCellEastSize = GetCellWidth(adaptRx);
-            grid.GridMaximumColumn = adaptRx.ColumnCount;
-            grid.GridMaximumRow = adaptRx.RowCount;
+            grid.GridMaximumColumn = (uint)adaptRx.ColumnCount;
+            grid.GridMaximumRow = (uint)adaptRx.RowCount;
             return grid;
         }
